InvokeWithException: make next successor a special begin node which has the same locationidentity as the invoke
InvokeWithExceptionNode has the unpleasant property that it is a ControlSplit
and MemoryCheckPoint at the same time.  In terms of scheduling, a node cannot
be placed after a ControlSplit node as it denotes the end of a block.  Thus,
instead of connecting a FloatingReadNode to the InvokeWithException node
directly (as lastLocationAccess), we point rather to the BeginNode
(non-exceptional case) or ExceptionBeginNode.

To preserve consistency regarding memory dependencies, the former node must be
also a MemoryCheckPoint.